cardinality stuffs

+++ optisolveapi/sat.py
@@ -82,109 +82,126 @@ class CNF(SolverBase):
         # b=1 => ab=1
         self.add_clause([-b, ab])
 
-    # def SeqInc(self, vec):
-    #     return [self.ONE] + list(vec)
+    def SeqInc(self, vec):
+        return [self.ONE] + list(vec)
 
-    # def SeqAddConst(self, vec, c):
-    #     return [self.ONE] * c + list(vec)
+    def SeqAddConst(self, vec, c):
+        return [self.ONE] * c + list(vec)
 
-    # def SeqAdd(self, vec1, vec2):
-    #     n1 = len(vec1)
-    #     n2 = len(vec2)
-    #     vec3 = [self.var() for i in range(n1 + n2)]
-    #     ands = {}
+    def SeqAdd(self, vec1, vec2):
+        n1 = len(vec1)
+        n2 = len(vec2)
+        vec3 = [self.var() for i in range(n1 + n2)]
+        ands = {}
 
-    #     # self.constraint_unary(vec1)  # optional
-    #     # self.constraint_unary(vec2)  # optional
-    #     self.constraint_unary(vec3)
+        # self.constraint_unary(vec1)  # optional
+        # self.constraint_unary(vec2)  # optional
+        self.constraint_unary(vec3)
 
-    #     for i in range(n1):
-    #         ands[i, -1] = vec1[i]
-    #         for j in range(n2):
-    #             ands[i, j] = self.var()
-    #             self.constraint_and(vec1[i], vec2[j], ands[i, j])
-    #             ands[-1, j] = vec2[j]
+        for i in range(n1):
+            ands[i, -1] = vec1[i]
+            for j in range(n2):
+                ands[i, j] = self.var()
+                self.constraint_and(vec1[i], vec2[j], ands[i, j])
+                ands[-1, j] = vec2[j]
 
-    #     for isum in range(1, n1+n2+1):
-    #         clause0 = [-vec3[isum-1]]
-    #         for i in range(min(isum + 1, n1 + 1)):
-    #             vi = vec1[i-1] if i else 0
-    #             j = isum - i
-    #             if j > n2:
-    #                 continue
-    #             vj = vec2[j-1] if j else 0
+        for isum in range(1, n1+n2+1):
+            clause0 = [-vec3[isum-1]]
+            for i in range(min(isum + 1, n1 + 1)):
+                vi = vec1[i-1] if i else 0
+                j = isum - i
+                if j > n2:
+                    continue
+                vj = vec2[j-1] if j else 0
 
-    #             # vec1[i] = 1, vec2[j] = 1 => vec3[i][isum] = 1
-    #             clause = [vec3[isum-1], -vi, -vj]
+                # vec1[i] = 1, vec2[j] = 1 => vec3[i][isum] = 1
+                clause = [vec3[isum-1], -vi, -vj]
 
-    #             clause = [c for c in clause if c]
-    #             self.add_clause(clause)
+                clause = [c for c in clause if c]
+                self.add_clause(clause)
 
-    #             clause0.append(ands[i-1, j-1])
+                clause0.append(ands[i-1, j-1])
 
-    #         # FORALL i, j vec1[i] & vec2[j] = 0 => vec3[i][isum] = 0
-    #         clause0 = [c for c in clause0 if c]
-    #         self.add_clause(clause0)
-    #     return vec3
+            # FORALL i, j vec1[i] & vec2[j] = 0 => vec3[i][isum] = 0
+            clause0 = [c for c in clause0 if c]
+            self.add_clause(clause0)
+        return vec3
 
-    # def SeqAddMany(self, *vecs):
-    #     lst = list(vecs)
-    #     while len(lst) >= 2:
-    #         lst2 = []
-    #         shuffle(lst)
-    #         while len(lst) >= 2:
-    #             lst2.append(self.SeqAdd(lst.pop(), lst.pop()))
-    #         if lst:
-    #             lst2.append(lst.pop())
-    #         lst = lst2
-    #     return lst[0]
+    def SeqAddMany(self, *vecs):
+        lst = list(vecs)
+        while len(lst) >= 2:
+            lst2 = []
+            shuffle(lst)
+            while len(lst) >= 2:
+                lst2.append(self.SeqAdd(lst.pop(), lst.pop()))
+            if lst:
+                lst2.append(lst.pop())
+            lst = lst2
+        return lst[0]
 
-    # def SeqEq(self, vec1, vec2):
-    #     if len(vec1) < len(vec2):
-    #         self.add_clause([-vec2[len(vec1)]])
-    #     elif len(vec2) < len(vec1):
-    #         self.add_clause([-vec1[len(vec2)]])
-    #     for a, b in zip(vec1, vec2):
-    #         self.add_clause([a, -b])
-    #         self.add_clause([-a, b])
+    def SeqEq(self, vec1, vec2):
+        if len(vec1) < len(vec2):
+            self.add_clause([-vec2[len(vec1)]])
+        elif len(vec2) < len(vec1):
+            self.add_clause([-vec1[len(vec2)]])
+        for a, b in zip(vec1, vec2):
+            self.add_clause([a, -b])
+            self.add_clause([-a, b])
 
-    # def SeqEqConst(self, vec, c):
-    #     assert 0 <= c <= len(vec)
-    #     if c == 0:
-    #         self.add_clause([-vec[0]])
-    #     elif c == len(vec):
-    #         self.add_clause([vec[-1]])
-    #     else:
-    #         self.add_clause(vec[c-1])
-    #         self.add_clause(-vec[c])
+    def SeqEqConst(self, vec, c):
+        assert 0 <= c <= len(vec)
+        if c == 0:
+            self.add_clause([-vec[0]])
+        elif c == len(vec):
+            self.add_clause([vec[-1]])
+        else:
+            self.add_clause(vec[c-1])
+            self.add_clause(-vec[c])
 
     def Cardinality(self, vec, lim=None, shuffle=False):
         """
         [Sinz2005]-like cardinality.
+        Returns:
+        [>=0, >=1, >=2, ..., >= len(vec)+1]
+        =
+        [TRUE, ..., ..., ..., FALSE]
+        if lim is less, then the last one may not be added
+        [>=0, >=1, >=2, ..., >= lim]
+
+        lim = 2
+        [0, 1, 2] l=3
+        lim = 3
+        [0, 1, 2, Z] l=4
         """
         if lim is None:
-            lim = len(vec)
+            nvars = len(vec)
+            lim = len(vec) + 1
         else:
-            lim = min(lim, len(vec))
+            nvars = min(lim, len(vec))
 
         assert vec
         if len(vec) == 1:
-            return list(vec)
+            res = [self.ONE, vec[0]]
+            res += [self.ZERO] * (lim + 1 - len(res))
+            return res
 
         if shuffle:
             vec = list(vec)
             _shuffle(vec)
 
         sub = self.Cardinality(vec[:-1], lim=lim, shuffle=False)
-        res = [self.var() for _ in range(lim)]
+        res = [self.ONE] + [self.var() for _ in range(nvars)]
+        res += [self.ZERO] * (lim + 1 - len(res))
         var = vec[-1]
 
-        # res[i] = card >= i+1
-        # res[0] = sub[0] | var
-        self.constraint_or(sub[0], var, res[0])
-        for i in range(1, lim):
+        # res[i] = card >= i
+
+        # res[0] = True
+        # res[1] = sub[1] | var
+        self.constraint_or(sub[1], var, res[1])
+        for i in range(2, nvars + 1):
             # res[i] = sub[i] | (sub[i-1] & var)
-            if len(sub) >= i + 1:
+            if len(sub) >= i + 1 and sub[i] != self.ZERO:
                 x0, x1, x2, x3 = sub[i], sub[i-1], var, res[i]
 
                 # Sinz claims that if we encode Cardinality <= k
@@ -198,7 +215,8 @@ class CNF(SolverBase):
                 self.add_clause([-x1, -x2, x3])
                 self.add_clause([-x0, x3])
             else:
-                assert i == lim - 1
+                assert i in (len(vec), lim)
+                assert sub[i] == self.ZERO
                 self.constraint_and(sub[i-1], var, res[i])
         return res
 
@@ -238,47 +256,47 @@ class CNF(SolverBase):
     #         S.add_clause([-vdst] + [vsrc for vsrc in sub])
     #     return dst
 
-    # def SeqMultConst(self, src, c):
-    #     res = []
-    #     for v in src:
-    #         res += [v] * c
-    #     return res
+    def SeqMultConst(self, src, c):
+        res = []
+        for v in src:
+            res += [v] * c
+        return res
 
-    # def AlignPad(self, a, b):
-    #     n = min(len(a), len(b)) + 1
-    #     a = list(a) + [self.ZERO] * (n - len(a))
-    #     b = list(b) + [self.ZERO] * (n - len(b))
-    #     return a, b
+    def AlignPad(self, a, b):
+        n = min(len(a), len(b)) + 1
+        a = list(a) + [self.ZERO] * (n - len(a))
+        b = list(b) + [self.ZERO] * (n - len(b))
+        return a, b
 
-    # def SeqLess(self, a, b):
-    #     # 1 0
-    #     # 0 0
-    #     a, b = self.AlignPad(a, b)
-    #     n = len(a)
+    def SeqLess(self, a, b):
+        # 1 0
+        # 0 0
+        a, b = self.AlignPad(a, b)
+        n = len(a)
 
-    #     # Bad (equal):
-    #     # 1 0
-    #     # 1 0
-    #     for i in range(n-1):
-    #         self.add_clause([-a[i], -b[i], a[i+1], b[i+1]])
+        # Bad (equal):
+        # 1 0
+        # 1 0
+        for i in range(n-1):
+            self.add_clause([-a[i], -b[i], a[i+1], b[i+1]])
 
-    #     # Bad (greater):
-    #     # 1
-    #     # 0
-    #     for i in range(n):
-    #         self.add_clause([-a[i], b[i]])
+        # Bad (greater):
+        # 1
+        # 0
+        for i in range(n):
+            self.add_clause([-a[i], b[i]])
 
-    # def SeqLessEqual(self, a, b):
-    #     # 1 0
-    #     # 0 0
-    #     a, b = self.AlignPad(a, b)
-    #     n = len(a)
+    def SeqLessEqual(self, a, b):
+        # 1 0
+        # 0 0
+        a, b = self.AlignPad(a, b)
+        n = len(a)
 
-    #     # Bad (greater):
-    #     # 1
-    #     # 0
-    #     for i in range(n):
-    #         self.add_clause([-a[i], b[i]])
+        # Bad (greater):
+        # 1
+        # 0
+        for i in range(n):
+            self.add_clause([-a[i], b[i]])
 
     def constraint_remove_lower(self, x: list, mx: int):
         clause = []
@@ -376,7 +394,9 @@ class PySAT(CNF):
         return tuple(sol[abs(v)] ^ (1 if v < 0 else 0) for v in vec)
 
     def __del__(self):
-        self._solver.delete()
+        if hasattr(self, "_solver"):
+            self._solver.delete()
+
 
 @CNF.register("pysat/formula")
 class Formula(CNF):
